Split basic structs and utility functions
