+++ wxChaos/core/FractalClasses.h
@@ -27,253 +27,20 @@ using namespace std;
 
 int Get_Cores();
 
-/****************************
-*                           *
-*      Fractal classes      *
-*                           *
-****************************/
-
-/**
-* @enum FractalType
-* @brief Enumerates the list of available fractals.
-*/
-enum class FractalType
-{
-    Mandelbrot,
-    MandelbrotZN,
-    Julia,
-    JuliaZN,
-    NewtonRaphsonMethod,
-    Sinoidal,
-    Magnetic,
-    Medusa,
-    Manowar,
-    ManowarJulia,
-    SierpinskyTriangle,
-    FixedPoint1,
-    FixedPoint2,
-    FixedPoint3,
-    FixedPoint4,
-    Tricorn,
-    BurningShip,
-    BurningShipJulia,
-    Fractory,
-    Cell,
-    DoublePendulum,
-    UserDefined,
-    FixedPointUserDefined,
-    ScriptFractal,
-    Undefined
-};
-
-/**
-* @enum Color
-* @brief Enumerates the rgb color.
-*/
-enum class Color
-{
-    Red,
-    Green,
-    Blue
-};
-
-/**
-* @enum RenderingAlgorithm
-* @brief List of available render algorithms.
-*/
-enum class RenderingAlgorithm
-{
-    EscapeTime,
-    GaussianInt,
-    EscapeAngle,
-    TriangleInequality,
-    ChaoticMap,
-    Lyapunov,
-    ConvergenceTest,
-    Other
-};
-
-/**
-* @enum ColorMode
-* @brief List of available color modes.
-*/
-enum class ColorMode
-{
-    Gradient,
-    Gaussian
-};
-
-/**
-* @enum Direction
-* @brief Movement direction.
-*/
-enum Direction
-{
-    Left = 0,
-    Right,
-    Up,
-    Down
-};
-
-/**
-* @struct LineData
-* @brief Parameters of line shape.
-*/
-struct LineData
-{
-    double x1, y1, x2, y2;
-    sf::Color color;
-};
-
-/**
-* @struct CircleData
-* @brief Parameters of circle shape.
-*/
-struct CircleData
-{
-    double x_center, y_center, radius;
-    sf::Color color;
-};
-
-/**
-* @struct Options
-* @brief Used to copy the parameters of the fractal.
-*/
-
-struct Options
-{
-    double minX;
-    double maxX;
-    double minY;
-    double maxY;
-    double xFactor;
-    double yFactor;
-    unsigned maxIter;
-    int changeGradient;
-    wxGradient gradient;
-    ColorMode colorPaletteMode;
-    RenderingAlgorithm alg;
-    FractalType type;
-    int paletteSize;
-    int gaussianPaletteSize;
-    int gradPaletteSize;
-    PanelOptions panelOpt;
-
-    double kReal;
-    double kImaginary;
-
-    bool orbitTrapMode;
-    bool colorSet;
-    bool colorMode;
-    bool buddhaMode;
-    bool smoothRender;
-    bool justLaunchThreads;
-    bool relativeColor;
-    int redInt, greenInt, blueInt;
-    double redMean, greenMean, blueMean;
-    double redStdDev, greenStdDev, blueStdDev;
-    sf::Color fSetColor;
-
-    int screenHeight;
-    int screenWidth;
-};
-
-/**
-* @enum FormulaType
-* @brief Type of user formula.
-*/
-enum class FormulaType
-{
-    Complex = 1,
-    FixedPoint
-};
-
-/**
-* @struct FormulaOpt
-* @brief User formula parameters.
-*/
-
-struct FormulaOpt
-{
-    wxString userFormula;        ///< String containing the user formula.
-    FormulaType type;            ///< Type of user formula (Complex or fixed point).
-    bool julia;                  ///< Inform if the fractal is a Julia variety. Only on complex type.
-    int bailout;                 ///< Bailout of the complex fractal.
-};
-
-/**
-* @struct Vector2Int
-* @brief Stores x and y coordinates.
-*/
-struct Vector2Int
-{
-    int x, y;
-};
-
-/**
-* @struct Vector2Double
-* @brief Stores x and y coordinates. This variant comes with vector arithmetic operators.
-*/
-
-struct Vector2Double
-{
-    double x, y;
-
-    Vector2Double();
-    Vector2Double(double _x, double _y);
-    Vector2Double(const Vector2Int& v);
-
-    Vector2Double operator-() const;
-    Vector2Double& operator+=(const Vector2Double& v);
-
-    Vector2Double& operator*=(const double t);
-    Vector2Double& operator/=(const double t);
-
-    double Length() const;
-    double SquaredLength() const;
-};
-
-inline Vector2Double operator+(const Vector2Double& u, const Vector2Double& v)
-{
-    return Vector2Double(u.x + v.x, u.y + v.y);
-}
-inline Vector2Double operator-(const Vector2Double& u, const Vector2Double& v)
-{
-    return Vector2Double(u.x - v.x, u.y - v.y);
-}
-inline Vector2Double operator*(const Vector2Double& u, const Vector2Double& v)
-{
-    return Vector2Double(u.x * v.x, u.y * v.y);
-}
-inline Vector2Double operator*(double t, const Vector2Double& v)
-{
-    return Vector2Double(t * v.x, t * v.y);
-}
-inline Vector2Double operator*(const Vector2Double& v, double t)
-{
-    return t * v;
-}
-inline Vector2Double operator/(Vector2Double v, double t)
-{
-    return (1.0 / t) * v;
-}
-
-
-/**
-* @struct Rect
-* @brief Stores rectangle coordinates.
-*/
-struct Rect
-{
-    double left, bottom, right, top;
-
-    Rect();
-    Rect(double _left, double _bottom, double _right, double _top);
-    Vector2Double GetLowerBound();
-    Vector2Double GetHigherBound();
-    void SetLowerBound(Vector2Double lb);
-    void SetHigherBound(Vector2Double hb);
-};
+// Aggregated fractal-related types
+#include "base/FractalType.h"
+#include "base/Color.h"
+#include "base/RenderingAlgorithm.h"
+#include "base/ColorMode.h"
+#include "base/Direction.h"
+#include "base/LineData.h"
+#include "base/CircleData.h"
+#include "base/Options.h"
+#include "base/FormulaType.h"
+#include "base/FormulaOpt.h"
+#include "base/Vector2Int.h"
+#include "base/Vector2Double.h"
+#include "base/Rect.h"
 
 /**
 * @class RenderFractal
