+++ wxChaos/core/SystemUtils.h
@@ -1,0 +1,3 @@
+#pragma once
+
+int Get_Cores();
