+++ wxChaos/core/base/CircleData.h
@@ -1,0 +1,8 @@
+#pragma once
+#include <SFML/Graphics.hpp>
+
+struct CircleData
+{
+    double x_center, y_center, radius;
+    sf::Color color;
+};
