+++ wxChaos/core/base/Color.h
@@ -1,0 +1,8 @@
+#pragma once
+
+enum class Color
+{
+    Red,
+    Green,
+    Blue
+};
