+++ wxChaos/core/base/ColorMode.h
@@ -1,0 +1,7 @@
+#pragma once
+
+enum class ColorMode
+{
+    Gradient,
+    Gaussian
+};
