+++ wxChaos/core/base/Direction.h
@@ -1,0 +1,9 @@
+#pragma once
+
+enum Direction
+{
+    Left = 0,
+    Right,
+    Up,
+    Down
+};
