+++ wxChaos/core/base/FormulaType.h
@@ -1,0 +1,7 @@
+#pragma once
+
+enum class FormulaType
+{
+    Complex = 1,
+    FixedPoint
+};
