+++ wxChaos/core/base/FractalType.h
@@ -1,0 +1,30 @@
+#pragma once
+
+enum class FractalType
+{
+    Mandelbrot,
+    MandelbrotZN,
+    Julia,
+    JuliaZN,
+    NewtonRaphsonMethod,
+    Sinoidal,
+    Magnetic,
+    Medusa,
+    Manowar,
+    ManowarJulia,
+    SierpinskyTriangle,
+    FixedPoint1,
+    FixedPoint2,
+    FixedPoint3,
+    FixedPoint4,
+    Tricorn,
+    BurningShip,
+    BurningShipJulia,
+    Fractory,
+    Cell,
+    DoublePendulum,
+    UserDefined,
+    FixedPointUserDefined,
+    ScriptFractal,
+    Undefined
+};
