+++ wxChaos/core/base/LineData.h
@@ -1,0 +1,8 @@
+#pragma once
+#include <SFML/Graphics.hpp>
+
+struct LineData
+{
+    double x1, y1, x2, y2;
+    sf::Color color;
+};
