+++ wxChaos/core/base/RenderingAlgorithm.h
@@ -1,0 +1,13 @@
+#pragma once
+
+enum class RenderingAlgorithm
+{
+    EscapeTime,
+    GaussianInt,
+    EscapeAngle,
+    TriangleInequality,
+    ChaoticMap,
+    Lyapunov,
+    ConvergenceTest,
+    Other
+};
