+++ wxChaos/core/base/Vector2Int.h
@@ -1,0 +1,6 @@
+#pragma once
+
+struct Vector2Int
+{
+    int x, y;
+};
